* master: (24 commits)
  Addressed PR feedback
  Added error handling
  Fixed all failing unit tests
  Fix unit test
  Fix lint issues
  Updates to make this generic
  Create new unique name validator that can be used in all situations.
  Change style of buttons
  Removed unused import, moves tslint:disable to top of file
  For 401 log in failures show bad username/password message
  Address PR feedback
  Refactor list data source to take a typed config object
  Fix lint error
  Lint fix
  Wire in Confirmation Dialog
  Fixes relating to padding changes - Had issues before with the form and ngIf, hence replaced with [hidden]. - This seems to have been resolved, so reverting back to ngIf - Solves issue with scss last-of-type selector from previous commit
  - Now only list that uses both card and table views is app wall - Remove associated card components - cf/org/space filters are now disabled to start with rather than hidden - Fixed right padding of last div in list header
  Fix initial change of list page size - Old code to avoid spam removed - We now take better care not to raise pagination change events given current values
  Fix linting
  Fix broken actions mapping
  ...

+++ src/app/features/applications/application/application-tabs-base/application-tabs-base.component.ts
@@ -9,6 +9,14 @@ import { Subscription } from 'rxjs/Rx';
 import { UpdateApplication, DeleteApplication } from '../../../../store/actions/application.actions';
 import { RouterNav } from '../../../../store/actions/router.actions';
 import { AppMetadataTypes, GetAppSummaryAction, GetAppStatsAction } from '../../../../store/actions/app-metadata.actions';
+import { ConfirmationDialogService, ConfirmationDialog } from '../../../../shared/components/confirmation-dialog.service';
+
+// Confirmation dialogs
+const appStopConfirmation = new ConfirmationDialog('Stop Application', 'Are you sure you want to stop this Application?', 'Stop');
+const appStartConfirmation = new ConfirmationDialog('Start Application', 'Are you sure you want to start this Application?', 'Start');
+
+// App delete will have a richer delete experience
+const appDeleteConfirmation = new ConfirmationDialog('Delete Application', 'Are you sure you want to delete this Application?', 'Delete');
 
 @Component({
   selector: 'app-application-tabs-base',
@@ -22,36 +30,23 @@ export class ApplicationTabsBaseComponent implements OnInit, OnDestroy {
     private router: Router,
     private applicationService: ApplicationService,
     private entityService: EntityService,
-    private store: Store<AppState>
+    private store: Store<AppState>,
+    private confirmDialog: ConfirmationDialogService
   ) { }
+
 
   public async: any;
 
-  sub: Subscription[] = [];
   isFetching$: Observable<boolean>;
   application;
   applicationActions$: Observable<string[]>;
 
+  summaryDataChanging$: Observable<boolean>;
   appSub$: Subscription;
   entityServiceAppRefresh$: Subscription;
   autoRefreshString = 'auto-refresh';
 
-  isEditSummary = false;
-
-  summaryExpanded = true;
-
-  summaryDataChanging$: Observable<boolean>;
-
   autoRefreshing$ = this.entityService.updatingSection$.map(update => update[this.autoRefreshString] || { busy: false });
-
-  appEdits: UpdateApplication;
-  appDefaultEdits: UpdateApplication = {
-    enable_ssh: false,
-    instances: 0,
-    memory: 0,
-    name: '',
-    environment_json: {}
-  };
 
   tabLinks = [
     { link: 'summary', label: 'Summary' },
@@ -62,25 +57,12 @@ export class ApplicationTabsBaseComponent implements OnInit, OnDestroy {
     { link: 'events', label: 'Events' },
   ];
 
-  startEdit() {
-    this.isEditSummary = true;
-    this.setAppDefaults();
-  }
-
-  endEdit() {
-    this.isEditSummary = false;
-  }
-
-  saveEdits() {
-    this.endEdit();
-    this.applicationService.updateApplication(this.appEdits, [AppMetadataTypes.SUMMARY]);
-  }
-
   stopApplication() {
-    this.endEdit();
-    const stoppedString = 'STOPPED';
-    this.applicationService.updateApplication({ state: stoppedString }, []);
-    this.pollEntityService('stopping', stoppedString).subscribe();
+    this.confirmDialog.open(appStopConfirmation, () => {
+      const stoppedString = 'STOPPED';
+      this.applicationService.updateApplication({ state: stoppedString }, []);
+      this.pollEntityService('stopping', stoppedString).subscribe();
+    });
   }
 
   pollEntityService(state, stateString) {
@@ -91,25 +73,22 @@ export class ApplicationTabsBaseComponent implements OnInit, OnDestroy {
   }
 
   startApplication() {
-    this.endEdit();
-    const startedString = 'STARTED';
-    this.applicationService.updateApplication({ state: startedString }, []);
-
-    this.pollEntityService('starting', startedString)
-      .delay(1)
-      .subscribe();
-  }
-
-  setAppDefaults() {
-    this.appEdits = { ... this.appDefaultEdits };
+    this.confirmDialog.open(appStartConfirmation, () => {
+      const startedString = 'STARTED';
+      this.applicationService.updateApplication({ state: startedString }, []);
+      this.pollEntityService('starting', startedString)
+        .delay(1)
+        .subscribe();
+      });
   }
 
   deleteApplication() {
-    this.store.dispatch(new DeleteApplication(this.applicationService.appGuid, this.applicationService.cfGuid));
+    this.confirmDialog.open(appDeleteConfirmation, () => {
+      this.store.dispatch(new DeleteApplication(this.applicationService.appGuid, this.applicationService.cfGuid));
+    });
   }
 
   ngOnInit() {
-
     const { cfGuid, appGuid } = this.applicationService;
     // Auto refresh
     this.entityServiceAppRefresh$ = this.entityService.poll(10000, this.autoRefreshString).do(() => {
@@ -125,7 +104,6 @@ export class ApplicationTabsBaseComponent implements OnInit, OnDestroy {
         this.store.dispatch(new RouterNav({ path: ['applications'] }));
       }
     });
-    this.setAppDefaults();
 
     this.isFetching$ = this.applicationService.isFetchingApp$;
 
@@ -147,28 +125,9 @@ export class ApplicationTabsBaseComponent implements OnInit, OnDestroy {
       }
       return isFetchingApp || isUpdating;
     });
-
-    this.sub.push(this.summaryDataChanging$
-      .filter((isChanging) => {
-        return !isChanging;
-      })
-      .mergeMap(_ => {
-        return this.applicationService.application$;
-      })
-      .subscribe((application: ApplicationData) => {
-        this.appDefaultEdits = {
-          name: application.app.entity.name,
-          instances: application.app.entity.instances,
-          memory: application.app.entity.memory,
-          enable_ssh: application.app.entity.enable_ssh,
-          environment_json: application.app.entity.environment_json
-        };
-      }));
   }
 
-
   ngOnDestroy() {
-    this.sub.forEach(subscription => subscription.unsubscribe());
     this.appSub$.unsubscribe();
     this.entityServiceAppRefresh$.unsubscribe();
   }
